Allow user/pass to be managed outside Puppet
The "value" field in ini_setting is optional. If it's undefined, it will
ensure the setting is there, but not modify the value. The benefit to
this type of behavior is that I can ensure there is a "password =" line
in the configuration, but allow the user to manually set it outside of
Puppet. Future Puppet runs will not modify the user-entered value.

With the default being a blank string, Puppet will wipe out the user's
changes every time Puppet runs.